+++ magick/annotate.c
@@ -1231,10 +1231,10 @@ static unsigned int RenderFreetype(Image *image,const DrawInfo *draw_info,
                     ((*p) < 127 ? OpaqueOpacity : TransparentOpacity);
                 fill_color=draw_info->fill;
                 if (pattern != (Image *) NULL)
-                  fill_color=AcquireOnePixel(pattern,
-                    (long) (point.x+x-pattern->tile_info.x) % pattern->columns,
-                    (long) (point.y+y-pattern->tile_info.y) % pattern->rows,
-                    &image->exception);
+                  fill_color=AcquireOnePixel(pattern,(long) ((unsigned long)
+                    (point.x+x-pattern->tile_info.x) % pattern->columns),(long)
+                    ((unsigned long) (point.y+y-pattern->tile_info.y) %
+                    pattern->rows),&image->exception);
                 opacity=(Quantum) ((unsigned long) ((MaxRGB-opacity)*
                   (MaxRGB-fill_color.opacity))/MaxRGB);
                 if (!active)
@@ -1539,9 +1539,10 @@ static unsigned int RenderPostscript(Image *image,const DrawInfo *draw_info,
         for (x=0; x < (long) annotate_image->columns; x++)
         {
           if (pattern != (Image *) NULL)
-            fill_color=AcquireOnePixel(pattern,
-              (x-pattern->tile_info.x) % pattern->columns,
-              (y-pattern->tile_info.y) % pattern->rows,&image->exception);
+            fill_color=AcquireOnePixel(pattern,(long) ((unsigned long)
+              (x-pattern->tile_info.x) % pattern->columns),(long)
+              ((unsigned long) (y-pattern->tile_info.y) % pattern->rows),
+              &image->exception);
           q->opacity=(Quantum) (MaxRGB-((unsigned long) ((MaxRGB-
             PixelIntensityToQuantum(q))*(MaxRGB-fill_color.opacity))/MaxRGB));
           q->red=fill_color.red;
