+++ magick/render.c
@@ -514,9 +514,10 @@ MagickExport unsigned int ColorFloodfillImage(Image *image,
         {
           if (floodplane[y*image->columns+x])
             {
-              color=AcquireOnePixel(pattern,
-                (x-pattern->tile_info.x) % pattern->columns,
-                (y-pattern->tile_info.y) % pattern->rows,&image->exception);
+              color=AcquireOnePixel(pattern,(long) ((unsigned long)
+                (x-pattern->tile_info.x) % pattern->columns),(long)
+                ((unsigned long) (y-pattern->tile_info.y) % pattern->rows),
+                &image->exception);
               if (!pattern->matte)
                 color.opacity=OpaqueOpacity;
               if (color.opacity != TransparentOpacity)
@@ -3693,8 +3694,9 @@ static unsigned int DrawPolygonPrimitive(Image *image,const DrawInfo *draw_info,
       pattern=draw_info->fill_pattern;
       if (pattern != (Image *) NULL)
         fill_color=AcquireOnePixel(pattern,
-          ((x-pattern->tile_info.x)) % pattern->columns,
-          ((y-pattern->tile_info.y)) % pattern->rows,&image->exception);
+          (long) ((unsigned long) (x-pattern->tile_info.x) % pattern->columns),
+          (long) ((unsigned long) (y-pattern->tile_info.y) % pattern->rows),
+          &image->exception);
       fill_opacity=MaxRGB-fill_opacity*(MaxRGB-fill_color.opacity);
       if (fill_opacity != TransparentOpacity)
         *q=AlphaComposite(&fill_color,fill_opacity,q,
@@ -3702,8 +3704,9 @@ static unsigned int DrawPolygonPrimitive(Image *image,const DrawInfo *draw_info,
       pattern=draw_info->stroke_pattern;
       if (pattern != (Image *) NULL)
         stroke_color=AcquireOnePixel(pattern,
-          ((x-pattern->tile_info.x)) % pattern->columns,
-          ((y-pattern->tile_info.y)) % pattern->rows,&image->exception);
+          (long) ((unsigned long) (x-pattern->tile_info.x) % pattern->columns),
+          (long) ((unsigned long) (y-pattern->tile_info.y) % pattern->rows),
+          &image->exception);
       stroke_opacity=MaxRGB-stroke_opacity*(MaxRGB-stroke_color.opacity);
       if (stroke_opacity != TransparentOpacity)
         *q=AlphaComposite(&stroke_color,stroke_opacity,q,
@@ -3925,9 +3928,10 @@ MagickExport unsigned int DrawPrimitive(Image *image,const DrawInfo *draw_info,
                 }
               if (pattern != (Image *) NULL)
                 {
-                  color=AcquireOnePixel(pattern,
-                    (x-pattern->tile_info.x) % pattern->columns,
-                    (y-pattern->tile_info.y) % pattern->rows,&image->exception);
+                  color=AcquireOnePixel(pattern,(long) ((unsigned long)
+                    (x-pattern->tile_info.x) % pattern->columns),(long)
+                    ((unsigned long) (y-pattern->tile_info.y) % pattern->rows),
+                    &image->exception);
                   if (!pattern->matte)
                     color.opacity=OpaqueOpacity;
                 }
